--Add verification to when user create account

+++ src/main/java/com/cours/clientleger/Model/AccessingDataJPA/InternautesRepository.java
@@ -4,7 +4,7 @@ import com.cours.clientleger.Model.Database.Internautes;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface InternautesRepository extends CrudRepository<Internautes, Integer> {
+public interface InternautesRepository extends CrudRepository<Internautes, Long> {
     Internautes getByLoginAndPassword(String login, String password);
 
     Boolean existsByLoginAndPassword(String login, String password);
